WIP: mpool, cache friendly allocator for constant sized elements
Having the old code for the mpool around, this is the initial commit importing mpool
code and make it useable for trealla.

This becomes an experiment to check if there are sufficient performance gains using the mpool
for the skiplist and cell allocator.

NOTE:
The code is currently licensed under GPLv2. When this experiment becomses successful and
merged into the trealla mainline it will be relicensed under MIT License.