mpool: freeing unused clusters
When a cluster becomes free its kept in a linger-cache.
Any cluster already cached there will be freed when it is really free (might become reused
meanwhile).

The reason for this linger-cache is to keep memory for a short time in case it gets reused
soon. Otherwise freeing/allocating on cluster boundaries would cause unnecessary malloc/free
of whole clusters.