new ensure() and message() macros replacing abort()
ensure(predicate, ...) is similar to assert(predicate). Predicate is true the program
continues. When it is false then a message with the source location and condition that failed
plus optional formatted information is printed on stderr and the program aborts.

These optional information is given as additional arguments, where the first must be a printf
style format string followed by the things to be printed.

When NDEBUG is defined at build time then the message printing is omitted but the predicate
check and abort() on failure is still enabled.

The message(fmt, ...) implements the informal message printing part as described above. It can
be used to print debug messages unconditionally. Defining NDEBUG disables this.